requirements.txt
fastapi

uvicorn

streamlit

openai

faiss-cpu

sentence-transformers

pymupdf

python-multipart

tqdm

pydantic